+++ src/Data/Fields/Schema/FieldMultiasset.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Storyblok\ManagementApi\Data\Fields\Schema;
+
+class FieldMultiasset extends FieldGeneric
+{
+    /**
+     * @return array<mixed>
+     */
+    public function filetypes(): array
+    {
+        return $this->getArray("filetypes");
+    }
+}